styled backgrounds and moved images
