+++ src/components/BackButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TouchableOpacity, StyleSheet, BackHandler, Platform, Linking } from "react-native";
+import { TouchableOpacity, StyleSheet, BackHandler, Platform} from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import { useNavigation } from "@react-navigation/native";
