+++ src/screens/onboarding_1/components/OnboardingBackground.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { StyleSheet } from "react-native";
 import { ImageBackground } from "expo-image";
+import { styles } from "@src/screens/onboarding_1/styles";
 
 interface OnboardingBackgroundProps {
   children: React.ReactNode;
@@ -11,18 +11,10 @@ export default function OnboardingBackground({ children }: OnboardingBackgroundP
     <ImageBackground
       source={require("@assets/images/onboarding_1-bg.png")}
       contentFit="fill"
-      style={styles.image}
+      style={styles.backgroundImage}
     >
       {children}
     </ImageBackground>
   );
 }
 
-const styles = StyleSheet.create({
-  image: {
-    flex: 1,
-    justifyContent: "center",
-    width: "100%",
-    height: "100%",
-  },
-});
