+++ src/utils/deviceUtils.ts
